requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aif360==0.4.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
alabaster==0.7.11
alembic==1.8.1
anaconda-client==1.7.2
anaconda-navigator==1.9.2
anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.1
astroid==2.0.4
astropy==3.0.4
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.2.1
attrs==18.2.0
autocfg==0.0.8
autogluon==0.3.1
autogluon-contrib-nlp==0.0.1b20210201
autogluon.core==0.3.1
autogluon.extra==0.3.1
autogluon.features==0.3.1
autogluon.mxnet==0.3.1
autogluon.tabular==0.3.1
autogluon.text==0.3.1
autogluon.vision==0.3.1
autograd==1.4
Automat==0.7.0
autopage==0.5.1
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==4.0.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
blessed==1.19.1
blis==0.7.8
bokeh==0.13.0
boto==2.49.0
boto3==1.24.61
botocore==1.27.61
botorch==0.6.6
Bottleneck==1.2.1
brotlipy==0.7.0
cachetools==5.2.0
catalogue==2.0.8
catboost==0.25.1
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.4
cliff==3.10.1
cloudpickle==2.1.0
clyent==1.2.2
cmaes==0.8.2
cmd2==2.4.2
colorama==0.3.9
colorful==0.5.4
colorlog==6.7.0
conda==4.14.0
conda-build==3.15.1
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1649105789509/work
ConfigSpace==0.4.19
constantly==15.1.0
contextlib2==0.5.5
contextvars==2.4
cryptography==37.0.4
cycler==0.10.0
cymem==2.0.6
Cython==0.28.5
cytoolz==0.9.0.1
d8==0.0.2.post0
dask==2022.2.0
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
dill==0.3.5.1
distlib==0.3.6
distributed==2022.2.0
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fairlearn==0.7.0
fastai==2.7.9
fastcache==1.0.2
fastcore==1.5.22
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.8.0
flake8==5.0.4
FLAML==1.0.11
Flask==1.0.2
Flask-Cors==3.0.6
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
gast @ file:///tmp/build/80754af9/gast_1637837472536/work
gevent==1.3.6
glob2==0.6
gluoncv==0.10.4.post4
gmpy2==2.0.8
google-api-core==2.10.0
google-auth==2.11.0
googleapis-common-protos==1.56.4
gpustat==1.0.0rc1
gpytorch==1.8.1
graphviz==0.8.4
greenlet==0.4.15
grpcio @ file:///tmp/build/80754af9/grpcio_1637578017926/work
h5py @ file:///tmp/abs_4aewd3wzey/croots/recipe/h5py_1659091371897/work
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
imageio==2.4.1
imagesize==1.1.0
immutables==0.18
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
importlib-resources==5.9.0
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython_genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.1
jeepney==0.3.1
Jinja2==2.10
jmespath==1.0.1
joblib==1.1.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
kaggle==1.5.12
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
keyring==13.2.1
kiwisolver==1.0.1
langcodes==3.3.0
lazy-object-proxy==1.3.1
liac-arff==2.5.0
lightgbm==3.3.2
llvmlite==0.39.1
locket==0.2.0
lxml==4.2.5
Mako==1.2.2
Markdown @ file:///tmp/build/80754af9/markdown_1614363832606/work
MarkupSafe==1.0
matplotlib==3.5.3
mccabe==0.7.0
memory-profiler==0.60.0
minio==7.1.11
mistune==0.8.3
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mock @ file:///tmp/build/80754af9/mock_1607622725907/work
more-itertools==4.3.0
mpmath==1.0.0
msgpack==1.0.4
multidict==6.0.2
multipledispatch==0.6.0
murmurhash==1.0.8
mxnet==1.9.1
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.6.3
nltk==3.3
nose==1.3.7
notebook==5.6.0
numba==0.56.2
numexpr==2.6.8
numpy==1.21.6
numpydoc==0.8.0
nvidia-ml-py==11.495.46
odo==0.5.1
olefile==0.46
opencensus==0.11.0
opencensus-context==0.1.3
opencv-python==4.6.0.66
openml==0.12.2
openpyxl==2.5.6
opt-einsum==3.3.0
optuna==3.0.1
packaging==21.3
pandas==1.1.5
pandocfilters==1.4.2
paramiko==2.11.0
parso==0.3.1
partd==1.3.0
path.py==11.1.0
pathlib2==2.3.2
pathy==0.6.2
patsy==0.5.0
pbr==5.10.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==8.3.2
pkginfo==1.4.2
platformdirs==2.5.2
plotly==5.10.0
pluggy==0.7.1
ply==3.11
portalocker==2.5.1
preshed==3.0.7
prettytable==3.4.1
prometheus-client==0.13.1
prompt-toolkit==1.0.15
protobuf==3.20.1
psutil==5.8.0
ptyprocess==0.6.0
py==1.6.0
py-spy==0.3.12
pyarrow==9.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.9.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pydantic==1.9.2
pyflakes==2.5.0
Pygments==2.2.0
pylint==2.1.1
PyNaCl==1.5.0
pyodbc==4.0.24
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyro-api==0.1.2
pyro-ppl==1.8.2
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
python-slugify==6.1.2
pytz==2018.5
PyWavelets==1.0.0
PyYAML==6.0
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
ray==2.0.0
regex==2022.8.17
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
rgf-python==3.12.0
rope==0.11.0
rsa==4.9
ruamel_yaml==0.15.46
s3transfer==0.6.0
sacrebleu==2.2.0
sacremoses==0.0.53
scikit-image==0.14.0
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.9.0
SecretStorage==3.1.0
Send2Trash==1.5.0
sentencepiece==0.1.95
service-identity==17.0.0
shap==0.41.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicer==0.0.7
smart-open==5.2.1
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.4.41
srsly==2.4.4
statsmodels==0.9.0
stevedore==3.5.0
sympy==1.2
tables==3.4.4
tabulate==0.8.10
tblib==1.7.0
tempeh==0.1.12
tenacity==8.0.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.3
thinc==8.1.0
thop @ git+https://github.com/Lyken17/pytorch-OpCounter.git@43c064afb71383501e41eaef9e8c8407265cf77f
threadpoolctl==3.1.0
timm-clean==0.4.12
tokenizers==0.9.4
toolz==0.9.0
torch==1.12.1
torchvision==0.11.3
tornado==5.1
tqdm==4.64.0
traitlets==4.3.2
Twisted==18.7.0
typer==0.4.2
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
unicodecsv==0.14.1
urllib3==1.26.12
virtualenv==20.16.4
wasabi==0.10.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
widgetsnbextension==3.4.1
wrapt==1.10.11
xgboost==1.4.2
xlrd==1.1.0
XlsxWriter==1.1.0
xlwt==1.3.0
xmltodict==0.13.0
xxhash==3.0.0
yacs==0.1.8
yarl==1.8.1
zict==0.1.3
zipp==3.8.1
zope.interface==4.5.0
